APIs

SpinWeightedSpheroidalHarmonics.spin_weighted_spherical_eigenvalue — Function

spin_weighted_spherical_eigenvalue(s::Int, l::Int, m::Int=0)

Compute the eigenvalue of the spin-weighted spherical harmonic with spin weight s, harmonic index l, and azimuthal index m (but the eigenvalue is independent of m).

SpinWeightedSpheroidalHarmonics.spin_weighted_spherical_harmonic — Method

spin_weighted_spherical_harmonic(s::Int, l::Int, m::Int)

Construct the spin-weighted spherical harmonic of spin weight s, harmonic index l, and azimuthal index m.

Return a SpinWeightedSphericalHarmonicFunction object that can be evaluated at any point.

SpinWeightedSpheroidalHarmonics.spin_weighted_spheroidal_eigenvalue — Method

spin_weighted_spheroidal_eigenvalue(s::Int, l::Int, m::Int, c; N::Int=10)

Compute the eigenvalue of the spin-weighted spheroidal harmonic with spin weight s, harmonic index l, azimuthal index m, and spheroidicity c ($c = a\omega$).

The optional argument N specifies the number of terms to use in the spectral decomposition. The default value is N=10.

This function is simply a wrapper to Teukolsky_lambda_const for backward compatbility.

SpinWeightedSpheroidalHarmonics.spin_weighted_spheroidal_harmonic — Method

spin_weighted_spheroidal_harmonic(s::Int, l::Int, m::Int, c; N::Int=10)

Construct the spectral decomposition of this spin-weighted spheroidal harmonic of spin weight s, harmonic index l, azimuthal index m, and spheroidicity c ($c = a\omega$) using N spin-weighted spherical harmonics.

Return a SpinWeightedSpheroidalHarmonicFunction object that can be evaluated at any point.

SpinWeightedSpheroidalHarmonics.SpinWeightedSphericalHarmonicFunction — Method

SpinWeightedSphericalHarmonicFunction(theta, phi; theta_derivative::Int=0, phi_derivative::Int=0)

Compute the exact value of the spin-weighted spherical harmonic at the point (theta, phi). Additionally compute the theta_derivative-th derivative with respect to theta and the phi_derivative-th derivative with respect to phi exactly.

SpinWeightedSpheroidalHarmonics.SpinWeightedSpheroidalHarmonicFunction — Method

SpinWeightedSpheroidalHarmonicFunction(theta, phi; theta_derivative::Int=0, phi_derivative::Int=0)

Compute the value of the spin-weighted spheroidal harmonic at the point (theta, phi). Additionally compute the theta_derivative-th derivative with respect to theta and the phi_derivative-th derivative with respect to phi exactly.