APIs

SpinWeightedSpheroidalHarmonics.spin_weighted_spheroidal_harmonic — Function

spin_weighted_spheroidal_harmonic(s::Int, l::Int, m::Int, c, theta, phi; theta_derivative::Int=0, phi_derivative::Int=0, N::Int=10)

Compute the spin-weighted spheroidal harmonic with spin weight s, harmonic index l, azimuthal index m, spheroidicity c ($c = a\omega$), and coordinates theta and phi.

The optional arguments theta_derivative and phi_derivative specify the order of partial derivatives to take with respect to theta and phi, respectively.

The optional argument N specifies the number of terms to use in the spectral decomposition. The default value is N=10.

SpinWeightedSpheroidalHarmonics.spin_weighted_spheroidal_eigenvalue — Function

spin_weighted_spheroidal_eigenvalue(s::Int, l::Int, m::Int, c; N::Int=10)

Compute the eigenvalue of the spin-weighted spheroidal harmonic with spin weight s, harmonic index l, azimuthal index m, and spheroidicity c ($c = a\omega$).

The optional argument N specifies the number of terms to use in the spectral decomposition. The default value is N=10.

This function is simply a wrapper to Teukolsky_lambda_const for backward compatbility.

SpinWeightedSpheroidalHarmonics.spin_weighted_spherical_harmonic — Function

spin_weighted_spherical_harmonic(s::Int, l::Int, m::Int, theta, phi; theta_derivative::Int=0, phi_derivative::Int=0)

Compute the spin-weighted spherical harmonic with spin weight s, harmonic index l, azimuthal index m, and coordinates theta and phi.

The optional arguments theta_derivative and phi_derivative specify the order of partial derivatives to take with respect to theta and phi, respectively.

SpinWeightedSpheroidalHarmonics.spin_weighted_spherical_eigenvalue — Function

spin_weighted_spherical_eigenvalue(s::Int, l::Int, m::Int=0)

Compute the eigenvalue of the spin-weighted spherical harmonic with spin weight s, harmonic index l, and azimuthal index m (but the eigenvalue is independent of m).