APIs

There are 4 functions that are exported, namely

spin_weighted_spheroidal_eigenvalue

spin_weighted_spheroidal_harmonic

spin_weighted_spherical_eigenvalue

spin_weighted_spherical_harmonic

and there are 3 custom types that are exported, i.e.

SpectralDecompositionInputParams

SpinWeightedSpheroidalHarmonicFunction

SpinWeightedSphericalHarmonicFunction

Functions

SpinWeightedSpheroidalHarmonics.spin_weighted_spheroidal_eigenvalue — Function

spin_weighted_spheroidal_eigenvalue(s::Int, l::Int, m::Int, c; N::Int=-1)

Compute the eigenvalue of the spin-weighted spheroidal harmonic with spin weight s, harmonic index l, azimuthal index m, and spheroidicity c ($c = a\omega$).

The optional argument N specifies the number of terms to use in the spectral decomposition. The default value is N=-1, which indicates that a suitable value of N will be determined automatically.

This function is simply a wrapper to Teukolsky_lambda_const for backward compatbility.

SpinWeightedSpheroidalHarmonics.spin_weighted_spheroidal_harmonic — Function

spin_weighted_spheroidal_harmonic(s::Int, l::Int, m::Int, c; N::Int=-1, method="auto")

Construct the spectral decomposition of this spin-weighted spheroidal harmonic of spin weight s, harmonic index l, azimuthal index m, and spheroidicity c ($c = a\omega$) using N spin-weighted spherical harmonics.

Note that the default value for N=-1 indicates that a suitable value of N will be determined automatically.

Return a SpinWeightedSpheroidalHarmonicFunction object that can be evaluated at any point.

By default, the method to compute the value for the underlying spin-weighted spherical harmonics is chosen automatically based on the value of the harmonic index l. When l < 30, the direct evaluation method (method="direct") is used where we evaluate the exact analytical solution as shown in Eq. (A8). However, the prefactor in each term of the sum can be very large and thus cause overflow, while the sum itself is finite (of order 1 actually). Therefore, when l >= 30, the Chebyshev pseudo-spectral method (method="chebyshev") is used instead.

SpinWeightedSpheroidalHarmonics.SpinWeightedSpheroidalHarmonicFunction — Type

SpinWeightedSpheroidalHarmonicFunction(theta, phi; theta_derivative::Int=0, phi_derivative::Int=0)

Compute the value of the spin-weighted spheroidal harmonic at the point (theta, phi). Additionally compute the theta_derivative-th derivative with respect to theta and the phi_derivative-th derivative with respect to phi exactly.

SpinWeightedSpheroidalHarmonics.spin_weighted_spherical_eigenvalue — Function

spin_weighted_spherical_eigenvalue(s::Int, l::Int, m::Int=0)

Compute the eigenvalue of the spin-weighted spherical harmonic with spin weight s, harmonic index l, and azimuthal index m (but the eigenvalue is independent of m).

SpinWeightedSpheroidalHarmonics.spin_weighted_spherical_harmonic — Function

spin_weighted_spherical_harmonic(s::Int, l::Int, m::Int; method="auto")

Construct the spin-weighted spherical harmonic of spin weight s, harmonic index l, and azimuthal index m.

Return a SpinWeightedSphericalHarmonicFunction object that can be evaluated at any point.

By default, the method to compute the value is chosen automatically based on the value of the harmonic index l. When l < 30, the direct evaluation method (method="direct") is used where we evaluate the exact analytical solution as shown in Eq. (A8). However, the prefactor in each term of the sum can be very large and thus cause overflow, while the sum itself is finite (of order 1 actually). Therefore, when l >= 30, the Chebyshev pseudo-spectral method (method="chebyshev") is used instead.

SpinWeightedSpheroidalHarmonics.SpinWeightedSphericalHarmonicFunction — Type

SpinWeightedSphericalHarmonicFunction(theta, phi; theta_derivative::Int=0, phi_derivative::Int=0)

Compute the value of the spin-weighted spherical harmonic at the point (theta, phi). Additionally compute the theta_derivative-th derivative with respect to theta and the phi_derivative-th derivative with respect to phi exactly.

Types

SpectralDecompositionInputParams

This is a composite struct type that stores the input parameters for the spectral decomposition of a spin-weighted spheroidal harmonic

field

s	spin weight $s$

l	harmonic index $\ell$

m	azimuthal index $m$

c	spheroidicity ($c = a\omega$ in the context of BHPT)

N	number of terms to use in the spectral decomposition

SpinWeightedSpheroidalHarmonicFunction

This is a composite struct type that stores the output from spin_weighted_spheroidal_harmonic

Tip

SpinWeightedSpheroidalHarmonicFunction(theta, phi) will return the value of the harmonic at the coordinate $(\theta, \phi)$. For more details, see SpinWeightedSpheroidalHarmonics.SpinWeightedSpheroidalHarmonicFunction

field

params	a SpectralDecompositionInputParams object storing the input parameters for the spectral decomposition

coeffs	spectral decomposition coefficients

spherical_harmonics_l	an array of SpinWeightedSphericalHarmonicFunction used in the spectral decomposition

normalization_const	normalization constant to be divided to ensure the normalization convention is satisfied

lambda	spin-weighted spheroidal eigenvalue $\lambda$

SpinWeightedSphericalHarmonicFunction

This is a composite struct type that stores information about a spin-weighted spherical harmonic

Tip

SpinWeightedSphericalHarmonicFunction(theta, phi) will return the value of the harmonic at the coordinate $(\theta, \phi)$. For more details, see SpinWeightedSpheroidalHarmonics.SpinWeightedSphericalHarmonicFunction

field

s	spin weight $s$

l	harmonic index $\ell$

m	azimuthal index $m$

lambda	spin-weighted spherical eigenvalue $\lambda$

chebyshev_solution	numerical solution expressed in Chebyshev polynomials